smart-meter/requirements.txt
astroid==2.3.3
attrs==19.3.0
Click==7.0
cycler==0.10.0
Flask==1.1.1
Flask-JSON==0.3.4
importlib-metadata==1.1.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.0.0
numpy==1.17.4
packaging==19.2
pandas==0.25.3
pluggy==0.13.1
py==1.8.0
pylint==2.4.4
pyparsing==2.4.5
pytest==5.3.1
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
typed-ast==1.4.0
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
